cambio resend verification a auth y agrego logs
