+++ src/app.js
@@ -6,6 +6,10 @@ const { errorHandler } = require('./middlewares/errorHandler');
 const app = express();
 
 app.use(express.json());
+app.use((req, res, next) => {
+  console.log(`[${new Date().toISOString()}] ${req.method} ${req.url}`);
+  next();
+});
 
 app.get('/health', (_req, res) => {
   res.json({ status: 'ok' });
