+++ src/modules/auth/auth.controller.js
@@ -1,4 +1,5 @@
 const { authService } = require('./auth.service');
+const { userService } = require('../users/user.service');
 
 const authController = {
   async login(req, res, next) {
@@ -55,6 +56,15 @@ const authController = {
       const userId = req.user.sub;
       const result = await authService.changePassword(userId, req.body);
       res.status(200).json(result);
+    } catch (err) {
+      next(err);
+    }
+  },
+
+  async resendVerification(req, res, next) {
+    try {
+      await userService.resendVerification(req.body.email);
+      res.status(200).json({ message: 'Email de verificación reenviado.' });
     } catch (err) {
       next(err);
     }
